Add a cookie banner
This commit adds a cookie banner, which I have pilfered from the one
used on GOV.UK Publishing.

I chose this one because it’s the one viewed by the most people, so any
recommendations we make about it would have the biggest value.

It’s also likely to be the starting point for what services decide to
copy, in the absence of a component in the Design System.

I personally want to test this because I have strong reckons about the
negative impact on usability that cookie banners have, especially when
services have seen a majority of users ignore the banner rather than
dismiss it, therefore leading to it appearing on every. Single. Page.
Load. I reckon there is a real trade off to be made between the
usability of a service and the value of setting non-essential cookies.
If this is the case then it would be good for us to have done the
research to help qualify this trade off.

I have chosen to link to the help pages on GOV.UK if the user clicks any
of the links in the banner. This does take people out of the prototype
and we could consider making these links non-functional. But I feel that
the risk of people getting waylaid is true using a real service as well,
and we should see if this happens to people using our fake service.