Add ‘Not completed’ tags to the task list
We want to see if this makes it clearer to people what they need to do
next when they land on the page.

Uses the ‘tag’ component recently-added to the Design System. I chose
the grey one because the Design System seems to suggest it as the
default, and it doesn’t have the same semantic associations as red,
green or yellow.